fixes user retention #56
