+++ arches_lingo/src/arches_lingo/components/login/LoginForm.vue
@@ -1,5 +1,5 @@
 <script setup lang="ts">
-import { inject, ref } from "vue";
+import { ref } from "vue";
 import { useRoute, useRouter } from "vue-router";
 import { useGettext } from "vue3-gettext";
 
@@ -8,11 +8,7 @@ import InputText from "primevue/inputtext";
 import Button from "primevue/button";
 
 import { login } from "@/arches_lingo/api.ts";
-import {
-    DEFAULT_ERROR_TOAST_LIFE,
-    ERROR,
-    userKey,
-} from "@/arches_lingo/constants.ts";
+import { DEFAULT_ERROR_TOAST_LIFE, ERROR } from "@/arches_lingo/constants.ts";
 import { routeNames } from "@/arches_lingo/routes.ts";
 
 const { $gettext } = useGettext();
@@ -20,17 +16,12 @@ const toast = useToast();
 const router = useRouter();
 const route = useRoute();
 
-const { setUser } = inject(userKey, {
-    user: ref(null),
-    setUser: () => {},
-});
 const username = ref();
 const password = ref();
 
 const submit = async () => {
     try {
-        const userToSet = await login(username.value, password.value);
-        setUser(userToSet);
+        await login(username.value, password.value);
         router.push(route.redirectedFrom || { name: routeNames.root });
     } catch (error) {
         toast.add({
