+++ arches_lingo/src/arches_lingo/components/user/UserInteraction.vue
@@ -1,5 +1,5 @@
 <script setup lang="ts">
-import { computed, inject, ref } from "vue";
+import { computed, inject } from "vue";
 import { useRouter } from "vue-router";
 import { useGettext } from "vue3-gettext";
 
@@ -14,19 +14,17 @@ import {
 } from "@/arches_lingo/constants.ts";
 import { routeNames } from "@/arches_lingo/routes.ts";
 
-const { user, setUser } = inject(userKey, {
-    user: ref(null),
-    setUser: () => {},
-});
+import type { UserRefAndSetter } from "@/arches_lingo/types.ts";
 
 const { $gettext } = useGettext();
 const toast = useToast();
 const router = useRouter();
 
+const { user } = inject(userKey) as UserRefAndSetter;
+
 const issueLogout = async () => {
     try {
         await logout();
-        setUser(null);
         router.push({ name: routeNames.login });
     } catch (error) {
         toast.add({
