+++ arches_lingo/src/arches_lingo/constants.ts
@@ -1,6 +1,6 @@
 import type { InjectionKey } from "vue";
 
-import type { UserRefAndSetter } from "@/arches_lingo/types";
+import type { UserRefAndSetter } from "@/arches_lingo/types.ts";
 
 export const ERROR = "error";
 export const DEFAULT_ERROR_TOAST_LIFE = 8000;
