+++ arches_lingo/src/arches_lingo/routes.ts
@@ -3,31 +3,46 @@ export const routes = [
         path: "/",
         name: "root",
         component: () => import("@/arches_lingo/pages/HomePage.vue"),
-        meta: { shouldShowNavigation: true },
+        meta: {
+            shouldShowNavigation: true,
+            requiresAuth: true,
+        },
     },
     {
         path: "/login/:next?",
         name: "login",
         component: () => import("@/arches_lingo/pages/LoginPage.vue"),
-        meta: { shouldShowNavigation: false },
+        meta: {
+            shouldShowNavigation: false,
+            requiresAuth: false,
+        },
     },
     {
         path: "/search",
         name: "search",
         component: () => import("@/arches_lingo/pages/BasicSearch.vue"),
-        meta: { shouldShowNavigation: true },
+        meta: {
+            shouldShowNavigation: true,
+            requiresAuth: true,
+        },
     },
     {
         path: "/advanced-search",
         name: "advanced-search",
         component: () => import("@/arches_lingo/pages/AdvancedSearch.vue"),
-        meta: { shouldShowNavigation: true },
+        meta: {
+            shouldShowNavigation: true,
+            requiresAuth: true,
+        },
     },
     {
         path: "/schemes",
         name: "schemes",
         component: () => import("@/arches_lingo/pages/SchemeList.vue"),
-        meta: { shouldShowNavigation: true },
+        meta: {
+            shouldShowNavigation: true,
+            requiresAuth: true,
+        },
     },
 ];
